move kqp related tests to kqp (#39260)
