+++ ydb/core/kqp/ut/channels/ya.make
@@ -6,14 +6,13 @@ SIZE(MEDIUM)
 REQUIREMENTS(cpu:2)
 
 SRCS(
-    dq_arrow_helpers_ut.cpp
-    dq_output_channel_ut.cpp
-    ut_helper.cpp
+    dq_channel_service_ut.cpp
 )
 
 PEERDIR(
     library/cpp/testing/unittest
     library/cpp/threading/local_executor
+    ydb/core/kqp/ut/common
     yql/essentials/public/udf/service/exception_policy
     yql/essentials/sql/pg_dummy
 )
