+++ ydb/library/yql/dq/comp_nodes/ut/join_perf/rss-measure.sh
