+++ ydb/library/yql/dq/comp_nodes/ut/ya.make
@@ -3,9 +3,9 @@ UNITTEST_FOR(ydb/library/yql/dq/comp_nodes)
 PEERDIR(
     ydb/library/yql/dq/comp_nodes
     ydb/library/yql/dq/comp_nodes/ut/utils
+    ydb/library/yql/dq/comp_nodes/ut/join_perf
     yql/essentials/public/udf/service/exception_policy
     yql/essentials/sql/pg_dummy
-    ydb/core/kqp/tools/join_perf
 
     library/cpp/testing/unittest
     library/cpp/dwarf_backtrace
